Modify return value for timestamp::to_msecond and timestamp::to_usecond into long to long long.

+++ centreon-clib/src/timestamp.cc
@@ -261,7 +261,7 @@ void timestamp::sub_usecond(long usecond) {
  *
  *  @return The time in milliseconds.
  */
-long timestamp::to_msecond() const throw () {
+long long timestamp::to_msecond() const throw () {
   return (_second * 1000 + _usecond / 1000);
 }
 
@@ -279,7 +279,7 @@ time_t timestamp::to_second() const throw () {
  *
  *  @return The time in microseconds.
  */
-long timestamp::to_usecond() const throw () {
+long long timestamp::to_usecond() const throw () {
   return (_second * 1000000 + _usecond);
 }
